Safari-611.1.10/LayoutTests/http/tests/security/contentSecurityPolicy/xsl-redirect-blocked-expected.txt

CONSOLE MESSAGE: Refused to load http://localhost:8000/security/contentSecurityPolicy/resources/alert-fail.xsl because it does not appear in the script-src directive of the Content Security Policy.
